+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.2.6
+
+[WOO-1434](https://resursbankplugins.atlassian.net/browse/WOO-1434) Switching between stores may cause payment method desynch
+
+# 1.2.5
+
+[WOO-1433](https://resursbankplugins.atlassian.net/browse/WOO-1433) Performance issues in cart and checkout
+
 # 1.2.4
 
 * Finnish translation hotfix.